files dumped through pickle (movies.pkl)
app.py is a streamlit app making the frontend of the website(movie recommender system). movies.pkl is dataframe with three columns , movie_id , title and ,tags which is dump through pickle and is loaded in the app.py file. main_logic.py is the code notebook same as the MovieRecommenderSystemProject.ipynb but it is a .py file